azure-function/requirements.txt
azure-functions
requests
pandas
numpy
scikit-learn
matplotlib
statsmodels